Move XML fields parsing into a dedicated module called from IdpData
